converting countdown to context
