+++ 02-ignite-timer/src/pages/Home/index.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { createContext, useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 
 import { zodResolver } from "@hookform/resolvers/zod";
@@ -35,10 +35,17 @@ interface ICycle {
   finishedDate?: Date;
 }
 
+interface ICyclesContextType {
+  activeCycle: ICycle | undefined;
+  activeCycleId: string | null;
+  markCurrentCyclesAsFinished: () => void;
+}
+
+export const CyclesContext = createContext({} as ICyclesContextType);
+
 export default function Home() {
   const [cycles, setCycles] = useState<ICycle[]>([]);
   const [activeCycleId, setActiveCycleId] = useState<string | null>(null);
-  const [amoundSecondsPassed, setAmoundSecondsPassed] = useState(0);
 
   const { register, handleSubmit, watch, reset } = useForm<NewCycleFormData>({
     resolver: zodResolver(newCycleFormValidationSchema),
@@ -50,57 +57,23 @@ export default function Home() {
 
   const activeCycle = cycles.find((cycle) => cycle.id === activeCycleId);
 
-  const totalSeconds = activeCycle ? activeCycle.minutesAmount * 60 : 0;
-  const currentSeconds = activeCycle ? totalSeconds - amoundSecondsPassed : 0;
+  // const task = watch("task");
+  // const isSubmitDisabled = !task;
 
-  const minutesAmount = Math.floor(currentSeconds / 60);
-  const secondsAmount = currentSeconds % 60;
-
-  const minutes = minutesAmount.toString().padStart(2, "0");
-  const seconds = secondsAmount.toString().padStart(2, "0");
-
-  const task = watch("task");
-  const isSubmitDisabled = !task;
-
-  useEffect(() => {
-    if (activeCycle) {
-      const interval = setInterval(() => {
-        const secondsDifference = differenceInSeconds(
-          new Date(),
-          activeCycle.startDate
-        );
-
-        if (secondsDifference >= totalSeconds) {
-          setCycles((state) =>
-            state.map((cycle) => {
-              if (cycle.id === activeCycleId) {
-                return {
-                  ...cycle,
-                  finishedDate: new Date(),
-                };
-              }
-
-              return cycle;
-            })
-          );
-
-          setAmoundSecondsPassed(totalSeconds);
-
-          clearInterval(interval);
-        } else {
-          setAmoundSecondsPassed(secondsDifference);
+  function markCurrentCyclesAsFinished() {
+    setCycles((state) =>
+      state.map((cycle) => {
+        if (cycle.id === activeCycleId) {
+          return {
+            ...cycle,
+            finishedDate: new Date(),
+          };
         }
-      }, 1000);
 
-      return () => clearInterval(interval);
-    }
-  }, [activeCycle, activeCycleId, totalSeconds]);
-
-  useEffect(() => {
-    if (activeCycle) {
-      document.title = `${minutes}:${seconds}`;
-    }
-  }, [activeCycle, minutes, seconds, task]);
+        return cycle;
+      })
+    );
+  }
 
   function handleInterruptCycle() {
     setCycles((state) =>
@@ -119,27 +92,31 @@ export default function Home() {
     setActiveCycleId(null);
   }
 
-  function handleCreateNewCycle(data: NewCycleFormData) {
-    const newCycle: ICycle = {
-      id: new Date().getTime().toString(),
-      task: data.task,
-      minutesAmount: data.minutesAmount,
-      startDate: new Date(),
-    };
+  // function handleCreateNewCycle(data: NewCycleFormData) {
+  //   const newCycle: ICycle = {
+  //     id: new Date().getTime().toString(),
+  //     task: data.task,
+  //     minutesAmount: data.minutesAmount,
+  //     startDate: new Date(),
+  //   };
 
-    setCycles((oldCyclesValues) => [...oldCyclesValues, newCycle]);
-    setActiveCycleId(newCycle.id);
-    setAmoundSecondsPassed(0);
+  //   setCycles((oldCyclesValues) => [...oldCyclesValues, newCycle]);
+  //   setActiveCycleId(newCycle.id);
+  //   setAmoundSecondsPassed(0);
 
-    reset();
-  }
+  //   reset();
+  // }
 
   return (
     <HomeContainer>
-      <form onSubmit={handleSubmit(handleCreateNewCycle)} action="">
-        <NewCycleForm />
+      <form /* onSubmit={handleSubmit(handleCreateNewCycle)} */ action="">
+        <CyclesContext.Provider
+          value={{ activeCycle, activeCycleId, markCurrentCyclesAsFinished }}
+        >
+          {/* <NewCycleForm /> */}
 
-        <Countdown />
+          <Countdown />
+        </CyclesContext.Provider>
 
         {activeCycle ? (
           <StopCountdownButton type="button" onClick={handleInterruptCycle}>
@@ -147,7 +124,7 @@ export default function Home() {
             Interromper
           </StopCountdownButton>
         ) : (
-          <StartCountdownButton type="submit" disabled={isSubmitDisabled}>
+          <StartCountdownButton type="submit" /* disabled={isSubmitDisabled} */>
             <Play size={24} />
             Começar
           </StartCountdownButton>
